+++ app/src/main/java/com/infomaniak/drive/data/sync/FileObserveService.kt
@@ -17,7 +17,6 @@
  */
 package com.infomaniak.drive.data.sync
 
-import android.app.PendingIntent
 import android.app.Service
 import android.content.Intent
 import android.database.ContentObserver
@@ -28,10 +27,9 @@ import android.provider.MediaStore
 import android.util.Log
 import com.infomaniak.drive.data.models.MediaFolder
 import com.infomaniak.drive.data.models.UploadFile
-import com.infomaniak.drive.ui.menu.settings.SyncSettingsActivity
+import com.infomaniak.drive.data.sync.UploadAdapter.Companion.showSyncConfigNotification
 import com.infomaniak.drive.utils.SyncUtils.isSyncActive
 import com.infomaniak.drive.utils.SyncUtils.syncDelayJob
-import com.infomaniak.drive.utils.SyncUtils.syncImmediately
 import kotlinx.coroutines.Job
 
 
@@ -53,16 +51,12 @@ class FileObserveService : Service() {
         val syncSetting = UploadFile.getAppSyncSettings()!!
 
         if (syncSetting.syncVideo) {
-            val externalContentUri = MediaStore.Video.Media.EXTERNAL_CONTENT_URI
-            val internalContentUri = MediaStore.Video.Media.EXTERNAL_CONTENT_URI
-            contentResolver.registerContentObserver(externalContentUri, true, tableObserver)
-            contentResolver.registerContentObserver(internalContentUri, true, tableObserver)
+            contentResolver.registerContentObserver(MediaStore.Video.Media.EXTERNAL_CONTENT_URI, true, tableObserver)
+            contentResolver.registerContentObserver(MediaStore.Video.Media.INTERNAL_CONTENT_URI, true, tableObserver)
         }
 
-        val externalContentUri = MediaStore.Images.Media.EXTERNAL_CONTENT_URI
-        val internalContentUri = MediaStore.Images.Media.EXTERNAL_CONTENT_URI
-        contentResolver.registerContentObserver(externalContentUri, true, tableObserver)
-        contentResolver.registerContentObserver(internalContentUri, true, tableObserver)
+        contentResolver.registerContentObserver(MediaStore.Images.Media.EXTERNAL_CONTENT_URI, true, tableObserver)
+        contentResolver.registerContentObserver(MediaStore.Images.Media.INTERNAL_CONTENT_URI, true, tableObserver)
     }
 
     override fun onDestroy() {
@@ -87,11 +81,11 @@ class FileObserveService : Service() {
             uri?.let {
                 if (!applicationContext.isSyncActive()) {
                     when {
-                        MediaFolder.getAllSyncedFoldersCount() > 0 ->{
+                        MediaFolder.getAllSyncedFoldersCount() > 0 -> {
                             syncJob?.cancel()
                             syncJob = applicationContext.syncDelayJob()
                         }
-                        else -> FileObserveServiceApi24.showSyncConfigNotification(baseContext)
+                        else -> showSyncConfigNotification(baseContext)
                     }
                 }
             }
